backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic[email]==2.7.1
pyjwt==2.8.0
bcrypt==4.1.3
aiokafka==0.11.0
python-dotenv==1.0.1
httpx==0.27.0
aiosmtplib==3.0.1
qrcode[pil]==7.4.2
stripe==9.9.0
